moved check package into collector folder
