+++ pkg/collector/check/check.go
