+++ pkg/collector/check/check_test.go
