+++ pkg/collector/loader/file_provider.go
@@ -5,7 +5,7 @@ import (
 	"io/ioutil"
 	"path/filepath"
 
-	"github.com/DataDog/datadog-agent/pkg/check"
+	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	"github.com/op/go-logging"
 
 	"gopkg.in/yaml.v2"
