+++ pkg/core/loader.go
@@ -3,7 +3,7 @@ package core
 import (
 	"fmt"
 
-	"github.com/DataDog/datadog-agent/pkg/check"
+	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	"github.com/op/go-logging"
 )
