+++ pkg/py/config.go
@@ -3,7 +3,7 @@ package py
 import (
 	"reflect"
 
-	"github.com/DataDog/datadog-agent/pkg/check"
+	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	"github.com/mitchellh/reflectwalk"
 	"github.com/sbinet/go-python"
 )
