+++ pkg/py/loader_test.go
@@ -3,7 +3,7 @@ package py
 import (
 	"testing"
 
-	"github.com/DataDog/datadog-agent/pkg/check"
+	"github.com/DataDog/datadog-agent/pkg/collector/check"
 	"github.com/sbinet/go-python"
 )
